split constants and created folders and blank files for subsystems - rn
