+++ src/main/java/org/team157/robot/subsystems/flywheel/Flywheel.java
@@ -1,0 +1,5 @@
+package org.team157.robot.subsystems.flywheel;
+
+public class Flywheel {
+
+}
