+++ src/main/java/org/team157/robot/subsystems/flywheel/FlywheelSystem.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.team157.robot.subsystems;
+package org.team157.robot.subsystems.flywheel;
 
 // import the stuff 
 import static edu.wpi.first.units.Units.Degrees;
@@ -19,9 +19,9 @@ import yams.motorcontrollers.SmartMotorController;
 
 import org.team157.robot.Constants;
 import org.team157.robot.Constants.FieldConstants;
-import org.team157.robot.Constants.FlywheelConstants;
 import org.team157.robot.Constants.HoodConstants;
 import org.team157.robot.Constants.TelemetryConstants;
+import org.team157.robot.subsystems.vision.VisionSystem;
 
 import edu.wpi.first.math.Pair;
 import edu.wpi.first.math.controller.SimpleMotorFeedforward;
