+++ src/main/java/org/team157/robot/subsystems/hopper/HopperIOTalonFX.java
@@ -1,0 +1,5 @@
+package org.team157.robot.subsystems.hopper;
+
+public class HopperIOTalonFX {
+
+}
