+++ src/main/java/org/team157/robot/subsystems/hopper/HopperSystem.java
@@ -2,13 +2,12 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.team157.robot.subsystems;
+package org.team157.robot.subsystems.hopper;
 
 import static edu.wpi.first.units.Units.Inches;
 import static edu.wpi.first.units.Units.Kilograms;
 
 import org.team157.robot.Constants;
-import org.team157.robot.Constants.HopperConstants;
 import org.team157.robot.Constants.TelemetryConstants;
 
 import com.ctre.phoenix6.hardware.TalonFX;
