+++ src/main/java/org/team157/robot/subsystems/uptake/UptakeIO.java
@@ -1,0 +1,5 @@
+package org.team157.robot.subsystems.uptake;
+
+public class UptakeIO {
+
+}
