+++ src/main/java/org/team157/robot/subsystems/uptake/UptakeIOTalonFX.java
@@ -1,0 +1,5 @@
+package org.team157.robot.subsystems.uptake;
+
+public class UptakeIOTalonFX {
+
+}
