+++ src/main/java/org/team157/robot/subsystems/uptake/UptakeSystem.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.team157.robot.subsystems;
+package org.team157.robot.subsystems.uptake;
 
 import static edu.wpi.first.units.Units.Inches;
 import static edu.wpi.first.units.Units.Kilograms;
@@ -10,7 +10,6 @@ import static edu.wpi.first.units.Units.RPM;
 
 import org.team157.robot.Constants;
 import org.team157.robot.Constants.TelemetryConstants;
-import org.team157.robot.Constants.UptakeConstants;
 import com.ctre.phoenix6.hardware.TalonFX;
 
 import edu.wpi.first.math.system.plant.DCMotor;
